requirements.txt
# Автоматически оптимизированный файл зависимостей
# Основные библиотеки
aiogram>=3.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# База данных векторов
langchain>=0.0.267
langchain-community>=0.0.3
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.0.2
chromadb>=0.4.18

# OpenAI (минимум 1.56.0 для Responses/Conversations API)
openai>=1.56.0

# Обработка документов
PyPDF2>=3.0.1
python-docx>=0.8.11
markdown>=3.0

# Google API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
